edge/cache-controller/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
confluent-kafka==2.8.0
grpcio==1.70.0
grpcio-tools==1.70.0
idna==3.10
kafka==1.3.5
minio==7.2.15
protobuf==5.29.3
pycparser==2.22
pycryptodome==3.21.0
requests==2.32.3
setuptools==75.8.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
prometheus-client=0.21.1